Integrate chat input
