+++ portfolio/src/components/home/home.tsx
@@ -1,10 +1,25 @@
 import { useState, useEffect } from 'react';
 import LanguageSwitch, { LangProps } from '../LanguageSwitch';
 import { ChatBox } from 'react-chatbox-component';
-import AiChatBox from './AiChatBox';
 
 import MessageFormProps from './module/message_form';
 import FirstReply from './module/first_reply';
+
+async function callSelectFunction(text: string): Promise<string | undefined> {
+  const url = process.env.REACT_APP_SELECT_FUNCTION_URL || '/selectFunction';
+  try {
+    const res = await fetch(url, {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ text })
+    });
+    const data = await res.json();
+    return data.function as string;
+  } catch (err) {
+    console.error('Failed to call selectFunction', err);
+    return undefined;
+  }
+}
 
 import 'react-chatbox-component/dist/style.css';
 import './home.css';
@@ -16,6 +31,34 @@ export default function Home(props: LangProps) {
 
     const user = {
         "uid" : "Guest"
+    }
+
+    const handleSendMessage = async (input: string) => {
+        if (!input.trim()) return;
+
+        const userMsg: MessageFormProps = {
+            text: input,
+            id: messages.length + 1,
+            sender: {
+                uid: 'Guest',
+                name: 'Guest',
+                avatar: 'https://www.w3schools.com/howto/img_avatar.png'
+            }
+        };
+        setMessages(prev => [...prev, userMsg]);
+
+        const func = await callSelectFunction(input);
+        const botText = func ? `Function selected: ${func}` : 'Failed to get response';
+        const botMsg: MessageFormProps = {
+            text: botText,
+            id: userMsg.id + 1,
+            sender: {
+                uid: 'Takanori Kotama',
+                name: 'Takanori Kotama',
+                avatar: `${process.env.PUBLIC_URL}/kotama_icon.jpg`
+            }
+        };
+        setMessages(prev => [...prev, botMsg]);
     }
 
     useEffect(() => {
@@ -38,9 +81,9 @@ export default function Home(props: LangProps) {
                 <ChatBox
                     messages={messages}
                     user={user}
+                    onSubmit={handleSendMessage}
                 />
             </div>
-            <AiChatBox messages={messages} setMessages={setMessages} />
         </div>
     );
 }
